Modified Deck class.
