+++ PlayingCards/klonedike.py
@@ -1,8 +1,6 @@
-import os
-import random
 import tkinter as tk
 
-from base import BaseBoard, BaseCard, CardFace
+from base import BaseBoard, BaseCard, CardFace, Deck
 from Globals import BOARD_WIDTH, BOARD_HEIGHT, CARD_ROOT, MOVE_SPEED
 
 
@@ -32,26 +30,18 @@ OPENEDSTOCK = 'openedstock'
 ACESTOCK = 'acestock'
 
 
-class Deck:
+class KlonedikeDeck(Deck):
 
-    def __init__(self, cards_dir):
-        self._deck = [card for card in self.get_cards(cards_dir)]
+    def __init__(self):
+        super().__init__()
+        self._deck = [card for card in self.get_cards()]
 
-    def get_cards(self, cards_dir):
-        for path in cards_dir.iterdir():
+    def get_cards(self):
+        for path in self.cards_dir.iterdir():
             name = path.stem
             mark, value = name.split('_')
             if not mark.startswith('jocker'):
                 yield CardFace(tk.PhotoImage(file=path), mark, int(value))
-
-    def __getitem__(self, position):
-        return self._deck[position]
-
-    def __len__(self):
-        return len(self._deck)
-
-    def shuffle(self):
-        random.shuffle(self._deck)
 
 
 class CardOnBoard(BaseCard):
@@ -99,7 +89,7 @@ class Board(BaseBoard):
         self.is_start_horizontal_move = False
         self.is_start_stock_back = False
         self.holder = self.get_image('holder')
-        self.deck = Deck(self.cards_dir)
+        self.deck = KlonedikeDeck()
 
     def new_game(self):
         self.delete('all')
